bug 1462475: Rename DocumentRendered to Redirect
Rename the files as well, hopefully in a way that makes it easier to
follow the history of the file.
