+++ kuma/scrape/sources/document_redirect.py
